Fix RSSI capnp field: use setRssi() in Cv2xParameters, not setPower()
- vanetza.capnp: add rssi @2 :Int16 to Cv2xParameters to match vanetza schema
- cv2x_rpc_server: call cv2xParams.setRssi() so vanetza reads RSSI via getCv2x().getRssi()
- rssi type: int8_t -> int16_t throughout wrapper and server
- Remove empty placeholder application files (not part of this package)

+++ autotalks_wrapper.cpp
@@ -232,7 +232,7 @@ bool AutotalksCV2X::transmit(const uint8_t* data, size_t length, uint8_t priorit
 }
 
 bool AutotalksCV2X::receive(uint8_t* buffer, size_t* buffer_size,
-                           uint64_t* timestamp, int8_t* rssi, uint32_t* src_l2id,
+                           uint64_t* timestamp, int16_t* rssi, uint32_t* src_l2id,
                            uint32_t timeout_ms)
 {
     if (!initialized_ || !rx_socket_) {
@@ -267,7 +267,7 @@ bool AutotalksCV2X::receive(uint8_t* buffer, size_t* buffer_size,
     if (rssi != nullptr) {
         // CV2X RSSI is in dbm8 format (dBm * 8)
         // Convert to dBm by dividing by 8
-        *rssi = static_cast<int8_t>(rx_params.rssi[0] / 8);
+        *rssi = static_cast<int16_t>(rx_params.rssi[0] / 8);
         std::cout << "[RX] RSSI: " << (int)*rssi << " dBm (raw: " 
                   << rx_params.rssi[0] << ")" << std::endl;
     }
